Adding Apache 2 license
This license is approved by General Counsel and is the acceptable
and license agreed upon license for contributes to the Samvera project.

As one of the deliverables of the grant is to share this work with a
community, and seek community adoption and contribution, we want to
ensure that we are taking steps towards a community to which we have
an ongoing relationship with.

See [Samvera Community Intellectual Property Licensing and Ownership][1]
for further details.

[1]:https://wiki.duraspace.org/display/samvera/Samvera+Community+Intellectual+Property+Licensing+and+Ownership